feat: update build and dev process
