+++ azure-pipelines.yml
@@ -24,5 +24,5 @@ steps:
     npm test
     npm run lint
     npm run cs-check
-    npm run dist
+    npm run build
   displayName: 'npm install and build'
