+++ packages/core/package.json
@@ -3,32 +3,32 @@
   "version": "2.0.0-alpha.2",
   "description": "A simple React component capable of building HTML forms out of a JSON schema.",
   "scripts": {
-    "build:lib": "rimraf lib && cross-env NODE_ENV=production babel -d lib/ src/",
-    "build:dist": "rimraf dist && cross-env NODE_ENV=production webpack --config webpack.config.dist.js",
-    "cs-check": "prettier -l \"{playground,src,test}/**/*.js\"",
-    "cs-format": "prettier \"{playground,src,test}/**/*.js\" --write",
-    "dist": "npm run build:lib && npm run build:dist",
-    "lint": "eslint src test playground",
+    "build": "npm run build:cjs && npm run build:es",
+    "build:cjs": "NODE_ENV=production BABEL_ENV=cjs babel --extensions \".js,.jsx,.ts,.tsx\" ./src --out-dir ./dist/cjs",
+    "build:es": "NODE_ENV=production BABEL_ENV=es babel --extensions \".js,.jsx,.ts,.tsx\" ./src --out-dir ./dist/es",
+    "cs-check": "prettier -l \"{src,test}/**/*.js\"",
+    "cs-format": "prettier \"{src,test}/**/*.js\" --write",
+    "lint": "eslint src test",
     "precommit": "lint-staged",
-    "publish-to-npm": "npm run dist && npm publish",
-    "preversion": "npm run build:playground && npm run dist && npm run cs-check && npm run lint",
-    "start": "node devServer.js",
+    "publish-to-npm": "npm run build && npm publish",
+    "preversion": "npm run build && npm run cs-check && npm run lint",
+    "start": "concurrently \"npm:build:* --watch\"",
     "tdd": "cross-env NODE_ENV=test mocha --require @babel/register --watch --require ./test/setup-jsdom.js test/**/*_test.js",
     "test": "cross-env NODE_ENV=test mocha --require @babel/register --require ./test/setup-jsdom.js test/**/*_test.js",
     "test-coverage": "cross-env NODE_ENV=test nyc --reporter=lcov mocha --require @babel/register --require ./test/setup-jsdom.js test/**/*_test.js",
     "test-debug": "cross-env NODE_ENV=test mocha --require @babel/register --require ./test/setup-jsdom.js --debug-brk --inspect test/Form_test.js"
   },
   "lint-staged": {
-    "{playground,src,test}/**/*.js": [
+    "{src,test}/**/*.js": [
       "eslint --fix",
       "prettier --write",
       "git add"
     ]
   },
-  "main": "lib/index.js",
+  "main": "dist/cjs/index.js",
+  "module": "dist/es/index.js",
   "files": [
-    "dist",
-    "lib"
+    "dist"
   ],
   "engineStrict": false,
   "engines": {
@@ -39,7 +39,6 @@
     "react": ">=16"
   },
   "dependencies": {
-    "@babel/runtime-corejs2": "^7.4.5",
     "ajv": "^6.7.0",
     "core-js": "^2.5.7",
     "json-schema-merge-allof": "^0.6.0",
@@ -55,17 +54,20 @@
     "@babel/core": "^7.4.5",
     "@babel/plugin-proposal-class-properties": "^7.4.4",
     "@babel/plugin-proposal-object-rest-spread": "^7.4.4",
+    "@babel/plugin-proposal-optional-chaining": "^7.8.3",
+    "@babel/plugin-transform-object-assign": "^7.8.3",
     "@babel/plugin-transform-react-jsx": "^7.3.0",
     "@babel/plugin-transform-runtime": "^7.4.4",
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
     "@babel/register": "^7.4.4",
-    "@rjsf/playground": "^2.0.0-alpha.2.2",
+    "@babel/runtime-corejs2": "^7.4.5",
     "atob": "^2.0.3",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.6",
     "chai": "^3.3.0",
     "codemirror": "^5.30.0",
+    "concurrently": "^5.1.0",
     "create-react-ref": "^0.1.0",
     "cross-env": "^2.0.1",
     "css-loader": "^0.23.1",
