--- packages/core/webpack.config.dev.js
@@ -1,58 +1,0 @@
-var path = require("path");
-var webpack = require("webpack");
-const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
-
-module.exports = {
-  mode: "development",
-  devtool: "source-map",
-  entry: [
-    "webpack-hot-middleware/client?reload=true",
-    "./playground/app"
-  ],
-  output: {
-    path: path.join(__dirname, "build"),
-    filename: "bundle.js",
-    publicPath: "/static/"
-  },
-  plugins: [
-    new MonacoWebpackPlugin({
-      languages: ['json']
-    }),
-    new webpack.HotModuleReplacementPlugin(),
-  ],
-  resolve: {
-    symlinks: false,
-    alias: {
-      react: path.resolve('./node_modules/react'),
-      'react-dom': path.resolve('./node_modules/react-dom')
-    }
-  },
-  module: {
-    rules: [
-      {
-        test: /\.jsx?$/,
-        use: [
-          "babel-loader",
-        ],
-        include: [
-          path.join(__dirname, "src"),
-          path.join(__dirname, "playground"),
-          path.join(__dirname, "node_modules", "codemirror", "mode", "javascript"),
-        ]
-      },
-      {
-        test: /\.css$/,
-        use: [
-          "style-loader",
-          "css-loader",
-        ],
-        include: [
-          path.join(__dirname, "css"),
-          path.join(__dirname, "playground"),
-          path.join(__dirname, "node_modules", "monaco-editor"),
-          path.join(__dirname, "node_modules", "@rjsf/playground", "node_modules", "monaco-editor"),
-        ],
-      },
-    ]
-  }
-};
