+++ packages/material-ui/.gitignore
@@ -1,0 +1,1 @@
+.rts2_**
