+++ packages/playground/webpack.config.dev.js
@@ -7,8 +7,7 @@ module.exports = {
   mode: "development",
   devtool: "source-map",
   entry: [
-    "webpack-hot-middleware/client?reload=true",
-    "./playground/app"
+    "./src/index"
   ],
   output: {
     path: path.join(__dirname, "build"),
